move swr related to modal to stores/modal file
