+++ packages/app/src/components/PageCreateModal.jsx
@@ -14,7 +14,7 @@ import AppContainer from '~/client/services/AppContainer';
 import { withUnstatedContainers } from './UnstatedUtils';
 import { toastError } from '~/client/util/apiNotification';
 
-import { usePageCreateModal } from '~/stores/ui';
+import { usePageCreateModal } from '~/stores/modal';
 
 import PagePathAutoComplete from './PagePathAutoComplete';
